requirements.txt
bs4
yaspin
pysocks
requests
streamlit
python-dotenv
langchain-openai
langchain-ollama
backports.tarfile
langchain-anthropic
langchain_community
langchain_google_genai
playwright

# MCP Support
mcp>=1.0.0
httpx>=0.24.0
anyio>=3.7.0
pydantic>=2.0.0
nest-asyncio>=1.5.0